basic maping and functions works,
need to fulfil all functions, and testing all functions

+++ hibernate/spring_hibernate_relation/src/main/java/com/junjunguo/shr/dao/LocationDao.java
@@ -12,5 +12,5 @@ import java.util.List;
 public interface LocationDao {
     List<Location> findAllLocations();
 
-    Location findById(int id);
+    Location findById(long id);
 }
