+++ hibernate/spring_hibernate_relation/src/main/java/com/junjunguo/shr/service/serviceImpl/TagServiceImpl.java
@@ -25,7 +25,7 @@ public class TagServiceImpl implements TagService {
         return tagDao.findByLabel(label);
     }
 
-    public Tag findByTagId(int id) {
+    public Tag findByTagId(long id) {
         return tagDao.findById(id);
     }
